# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ДоВуз-Портал API
  version: 0.1.0
  description: |
    Контракт бэкенда АИС «ДоВуз-Портал». Снят со слоя `frontend/src/app/api/`
    и повторяет его как есть, включая смешанное именование полей: часть ответов
    в snake_case, часть в camelCase, query-параметры в camelCase.

    Аутентификация — HTTP-only кука сессии, которую выставляют `/auth/login`
    и `/auth/register`. Ошибки всегда приходят в формате `{"message": "..."}`.
    401 — «не вошёл» (фронт разлогинивает), 403 — «вошёл, но нельзя».
servers:
  - url: /v1
    description: Относительно хоста, с которого отдана спека; локально — http://localhost:8080/v1
security:
  - sessionCookie: []
...
paths:
...
  /accounts:
    get:
      tags: [accounts]
      operationId: getAllUsers
      summary: Все пользователи
      description: Только для администратора.
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Страница пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [accounts]
      operationId: saveInfo
      summary: Правка своего профиля
      description: |
        Меняются только переданные поля. Школа и класс проверяются на
        согласованность: класс должен принадлежать выбранной школе.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                nickname:
                  $ref: '#/components/schemas/Nickname'
                firstName:
                  type: string
                  maxLength: 100
                lastName:
                  type: string
                  maxLength: 100
                middleName:
                  type: string
                  maxLength: 100
                birthDate:
                  type: string
                  format: date
                description:
                  type: string
                  maxLength: 2000
                schoolId:
                  type: integer
                classId:
                  type: integer
                photo:
                  type: string
                  format: binary
                  description: Изображение JPEG, PNG, GIF или WebP, до 10 МБ
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
  /accounts/{account}:
    get:
      tags: [accounts]
      operationId: getAccount
      summary: Профиль по никнейму
      description: Email виден владельцу профиля, модераторам и администратору, остальным — `null`.
      parameters:
        - name: account
          in: path
          required: true
          description: Никнейм
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [accounts]
      operationId: deleteUser
      summary: Удалить пользователя
      description: Только для администратора, себя удалить нельзя.
      parameters:
        - name: account
          in: path
          required: true
          description: Идентификатор учётной записи
          schema:
            type: integer
      responses:
        '204':
          description: Удалён
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/{id}/ban:
    post:
      tags: [accounts]
      operationId: banUser
      summary: Заблокировать
      description: Закрывает все сессии пользователя. Только для администратора.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Новый статус
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusAccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/{id}/unban:
    post:
      tags: [accounts]
      operationId: unbanUser
      summary: Разблокировать
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Новый статус
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusAccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/{id}/profile:
    patch:
      tags: [accounts]
      operationId: adminUpdateUser
      summary: Правка профиля администратором
      description: |
        Меняются только переданные поля. `null` в `school_id` отвязывает от школы
        и класса, `null` в `class_id` — только от класса.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateProfileRequest'
      responses:
        '200':
          description: Обновлённый пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/{id}/role:
    put:
      tags: [accounts]
      operationId: updateRole
      summary: Сменить роль
      description: |
        Только для администратора, свою роль сменить нельзя. При снятии роли
        модератора снимаются и его назначения на предметы.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Новая роль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountUpdateRoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/{nickname}/course-progress:
    get:
      tags: [accounts]
      operationId: getCourseProgress
      summary: Прогресс по курсам
      description: По каждому курсу пользователя — занятия, баллы и статусы слушаний.
      parameters:
        - name: nickname
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Прогресс по каждому курсу
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseProgressResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/school/{schoolId}:
    get:
      tags: [accounts]
      operationId: getBySchool
      summary: Пользователи школы
      description: Для модераторов и администратора.
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Страница пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/class/{classId}:
    get:
      tags: [accounts]
      operationId: getByClass
      summary: Пользователи класса
      description: Модераторам и администратору — любой класс, ученику — только свой.
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Страница пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/me/password:
    put:
      tags: [accounts]
      operationId: changePassword
      summary: Сменить пароль
      description: Остальные сессии пользователя закрываются, текущая остаётся.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Пароль изменён
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /accounts/admin/register:
    post:
      tags: [accounts]
      operationId: adminRegisterUser
      summary: Создать пользователя администратором
      description: Без `role_id` создаётся «Пользователь».
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRegisterRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dovuz_session
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageNumber:
      name: pageNumber
      in: query
      description: Номер страницы, с нуля
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
        default: 0
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 10
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Нужно войти в систему
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: Файл слишком большой
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RoleTitle:
      type: string
      description: Роль так, как её показывает и сравнивает фронт
      enum: [Администратор, Модератор, Пользователь]
      x-enum-varnames: [ADMIN, MODERATOR, USER]
...
    Email:
      type: string
      description: Непустые части до и после «@», в домене есть точка; без пробелов и управляющих символов
      maxLength: 254
      pattern: '^[^@\x00-\x20\x7f]+@[^@\x00-\x20\x7f]+\.[^@\x00-\x20\x7f]+$'
    Nickname:
      type: string
      description: 3–32 символа — буквы, цифры, «_», «.», «-»; начинается с буквы или цифры
      pattern: '^[A-Za-zА-Яа-яЁё0-9][A-Za-zА-Яа-яЁё0-9_.-]{2,31}$'
    Password:
      type: string
      description: 8–72 символа, хотя бы одна заглавная буква, одна строчная и одна цифра
      minLength: 8
      maxLength: 72
...
    PageMeta:
      type: object
      required: [total_pages, total_size, page_number, page_size]
      properties:
        total_pages:
          type: integer
        total_size:
          type: integer
        page_number:
          type: integer
        page_size:
          type: integer
...
    AccountResponse:
      type: object
      required:
        - id
        - nickname
        - email
        - firstName
        - lastName
        - middleName
        - birthDate
        - description
        - isBanned
        - role
        - photoNameInDirectory
        - schoolId
        - schoolName
        - classId
        - className
      properties:
        id:
          type: integer
        nickname:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        middleName:
          type: [string, 'null']
        birthDate:
          type: [string, 'null']
          format: date
        description:
          type: [string, 'null']
        isBanned:
          type: boolean
        role:
          $ref: '#/components/schemas/RoleTitle'
        photoNameInDirectory:
          type: [string, 'null']
        schoolId:
          type: [integer, 'null']
        schoolName:
          type: [string, 'null']
        classId:
          type: [integer, 'null']
        className:
          type: [string, 'null']
    GetAllUserResponse:
      type: object
      required:
        - id
        - nickname
        - email
        - first_name
        - last_name
        - middle_name
        - photo_name_in_directory
        - school_id
        - school_name
        - class_id
        - class_name
        - role
        - is_banned
        - is_lagging
        - created_at
      properties:
        id:
          type: integer
        nickname:
          type: string
        email:
          type: [string, 'null']
        first_name:
          type: [string, 'null']
        last_name:
          type: [string, 'null']
        middle_name:
          type: [string, 'null']
        photo_name_in_directory:
          type: [string, 'null']
        school_id:
          type: [integer, 'null']
        school_name:
          type: [string, 'null']
        class_id:
          type: [integer, 'null']
        class_name:
          type: [string, 'null']
        role:
          $ref: '#/components/schemas/RoleTitle'
        is_banned:
          type: boolean
        is_lagging:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/GetAllUserResponse'
    StatusAccountResponse:
      type: object
      required: [id, is_banned]
      properties:
        id:
          type: integer
        is_banned:
          type: boolean
    UpdateRoleRequest:
      type: object
      required: [role_id]
      properties:
        role_id:
          type: integer
    AccountUpdateRoleResponse:
      type: object
      required: [id, role]
      properties:
        id:
          type: integer
        role:
          $ref: '#/components/schemas/RoleTitle'
    AdminUpdateProfileRequest:
      type: object
      properties:
        first_name:
          type: [string, 'null']
          maxLength: 100
        last_name:
          type: [string, 'null']
          maxLength: 100
        middle_name:
          type: [string, 'null']
          maxLength: 100
        school_id:
          type: [integer, 'null']
        class_id:
          type: [integer, 'null']
    AdminRegisterRequest:
      type: object
      required: [email, nickname, password]
      properties:
        email:
          $ref: '#/components/schemas/Email'
        nickname:
          $ref: '#/components/schemas/Nickname'
        password:
          $ref: '#/components/schemas/Password'
        role_id:
          type: [integer, 'null']
          description: Без поля или `null` — «Пользователь»
        school_id:
          type: [integer, 'null']
        class_id:
          type: [integer, 'null']
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          minLength: 1
          maxLength: 128
        new_password:
          $ref: '#/components/schemas/Password'
    HearingStageStatus:
      type: object
      required: [stage, status]
      properties:
        stage:
          type: string
        status:
          type: string
    CourseProgressResponse:
      type: object
      required:
        - course_id
        - course_name
        - total_lessons
        - submitted_lessons
        - accepted_lessons
        - average_score
        - max_possible_score
        - hearing_statuses
      properties:
        course_id:
          type: integer
        course_name:
          type: string
        total_lessons:
          type: integer
        submitted_lessons:
          type: integer
        accepted_lessons:
          type: integer
        average_score:
          type: number
        max_possible_score:
          type: number
        hearing_statuses:
          type: array
          items:
            $ref: '#/components/schemas/HearingStageStatus'
...